Make ">> user@team" work
All the logic for splitting out teams ran only if the operator was >>.

I refactored the method to extract the teamname by checking if we had a
teamname before checking what the operator was.

There is probably a more obvious/better way to do this (e.g in order to
get it to work I had to declare variables before I knew if I needed
them).