labels: lookup label by team
Because several teams can have the same label, when adding a label to a
task we need to get the label id that corresponds to the label on the
specific team we're working with.

Instead of doing a lookup for all labels, the teams lookup now returns
each teams labels, and we have a subroutine that returns a label id
given a label name and a team name.