Add cached attribute for project lookup
Returns a hash containing the name, id, description, teams, and issues
associated with the project

Projects are grouped by identifier, which is a #key at the beginning of
the project description:

```
{
  jorts => [ $p1 ],
  jean  => [ $p3 ],
}
```
To detect if two projects accidentally have the same identifier, the
returned list of a project lookup should have only one element.
Otherwise the calling code should return an error.