Be safer about removing http client on DEMOLISH
If you use Linear::Client in a program without a persistent loop, you
get warnings like this during global destruction:

(in cleanup) Notifier does not exist in collection at .../Linear/Client.pm line 57.

Using `remove_from_parent` is safer here, because if it doesn't have a
parent, it's a no-op.

(I've also added a predicate for _http, just so that we don't initialize
it _during_ demolish time, which would be silly.)