Add GitLab link as attachment to Linear issue
If the plan parser detects an MR in an issue title or description, it
will add it as an attachment to the issue.

To do that we first need to create the issue, then use the returned
issueId to do a createAttachment call with the attachment details.

We might also want to do a GitLab API call to get the commit message of
the MR so the link title is a little more descriptive than `hm!xxxx`