cached attrs: read the cache, don't regenerate it
What a goofy bug.

$self->$cache_attr_name is actually the builder.  It's badly named.
build the cache entry that we use to store the maybe-incomplete Future
that exists to prevent multi-fetch races (and to cache the known
values!)  We call it to initialize the cache if the cache isn't there at
all.

Accidentally, though, we were also calling that method to just see if
there was a cache.  This would never work, but we didn't inspect the
return type of the builder.  It would hand us a Future and we treated it
like the {time=>T, value=>V} cache entry.  Then we would think it was an
expired cache entry, because it has no data in it, and we'd throw it out
and call the builder *again* to populate the cache.  So we'd start a
fetch, throw it away because we didn't know it was a fetch, and then
start a NEW fetch which we would use.

Here, we get the cache entry from the cache, not the builder.  Now it
all works!