team macros: allow a team name to mean something more
For example, our BAK team is being retired, merged into a generic
product backlog.  We want people who try to send work to BAK to actually
have that work go to the generic backlog, but have the "Backend" guild
label added.

This is a bit of a bodge, but worked, and was fast.